fix: GA 코드 수정 (#96)
