+++ apps/web/src/app/layout.tsx
@@ -7,7 +7,6 @@ import Script from 'next/script';
 import './globals.css';
 import { Providers } from '@web/components/providers/Providers';
 import { META } from '@web/constants/metadata';
-import GoogleAnalytics from '@web/components/GoogleAnalytics/GoogleAnalytics';
 
 const pretendard = localFont({
   src: '../../public/font/subset-PretendardVariable.woff2',
@@ -34,12 +33,31 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const gaId = process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS;
+
   return (
     <html lang="ko">
       <body className={`${pretendard.className} ${themeClass}`}>
-        {process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS ? (
-          <GoogleAnalytics gaId={process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS} />
-        ) : null}
+        {gaId && (
+          <>
+            <Script
+              strategy="afterInteractive"
+              src={`https://www.googletagmanager.com/gtag/js?id=${gaId}`}
+            />
+            <Script
+              id="google-analytics"
+              strategy="afterInteractive"
+              dangerouslySetInnerHTML={{
+                __html: `
+                  window.dataLayer = window.dataLayer || [];
+                  function gtag(){dataLayer.push(arguments);}
+                  gtag('js', new Date());
+                  gtag('config', '${gaId}');
+                `,
+              }}
+            />
+          </>
+        )}
         <Script
           id="microsoft-clarity"
           strategy="afterInteractive"
